+++ src/schema/index.ts
@@ -189,6 +189,7 @@ export interface OAuth2AdditionalParameters {
 export interface OAuth2Auth {
   type: "oauth2"
   grant_type: OAuth2GrantType
+  discovery_url: string
   authorization_url: string
   access_token_url: string
   refresh_token_url: string